requirements.txt
discord.py
SQLAlchemy
python-dotenv
humanize